# cmd/executor/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateTodoItemRequest:
    properties:
      description:
        type: string
      dueDate:
        type: string
    required:
    - description
    - dueDate
    type: object
  dto.TodoItem:
    properties:
      createdAt:
        type: string
      description:
        type: string
      dueDate:
        type: string
      id:
        type: string
      updatedAt:
        type: string
    type: object
  dto.UpdateTodoItemRequest:
    properties:
      description:
        type: string
      dueDate:
        type: string
    required:
    - description
    - dueDate
    type: object
  response.ErrSwaggerResponse:
    properties:
      meta:
        properties:
          causes:
            items: {}
            type: array
          code:
            type: integer
          message:
            type: string
        type: object
      payload:
        additionalProperties: true
        type: object
    type: object
  response.ErrValidationSwaggerResponse:
    properties:
      meta:
        properties:
          causes:
            items:
              properties:
                field:
                  type: string
                message:
                  type: string
              type: object
            type: array
          code:
            type: integer
          message:
            type: string
        type: object
      payload:
        additionalProperties: true
        type: object
    type: object
info:
  contact:
    name: TodoAPP
    url: https://swagger.io/support
  termsOfService: http://swagger.io/terms/
paths:
  /todo-items:
    post:
      consumes:
      - application/json
      description: This api for create poll
      parameters:
      - description: Contains information to set data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTodoItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TodoItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrValidationSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
      security:
      - Bearer: []
      summary: Create TodoItem
      tags:
      - todo-items
  /todo-items/{id}:
    delete:
      consumes:
      - application/json
      description: This api for delete poll
      parameters:
      - description: TodoItem Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrValidationSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
      security:
      - Bearer: []
      summary: Delete TodoItem
      tags:
      - todo-items
    get:
      consumes:
      - application/json
      description: This api for poll by id
      parameters:
      - description: TodoItem Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TodoItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrValidationSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
      security:
      - Bearer: []
      summary: Get TodoItem By Id
      tags:
      - todo-items
    put:
      consumes:
      - application/json
      description: This api for update poll
      parameters:
      - description: TodoItem Id
        in: path
        name: id
        required: true
        type: string
      - description: Contains information to set data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTodoItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TodoItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrValidationSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
      security:
      - Bearer: []
      summary: Update TodoItem
      tags:
      - todo-items
  /todo-items/purge/{id}:
    delete:
      consumes:
      - application/json
      description: This api for purge poll
      parameters:
      - description: TodoItem Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.ErrValidationSwaggerResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrSwaggerResponse'
      security:
      - Bearer: []
      summary: Purge TodoItem
      tags:
      - todo-items
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the Authorization Bearer"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"